requirements.txt
streamlit
ultralytics
opencv-python-headless
torch
numpy
Pillow
